Hertta Advanced Job System
- Hertta can now process more than one job at a time
- Hertta now remembers the outcome of more than just the latest job
- Added weather forecast job
