+++ src/graphql.rs
@@ -13,7 +13,9 @@ mod state_input;
 mod time_line_input;
 mod topology_input;
 
-use crate::event_loop::{OptimizationState, OptimizationTask};
+use crate::event_loop;
+use crate::event_loop::job_store::JobStore;
+use crate::event_loop::jobs::{self, Job, JobOutcome, JobStatus, NewJob};
 use crate::input_data::Name;
 use crate::input_data_base::{
     BaseGenConstraint, BaseInputData, BaseMarket, BaseNode, BaseNodeDiffusion, BaseProcess,
@@ -24,7 +26,6 @@ use crate::scenarios::Scenario;
 use crate::settings::{LocationSettings, Settings};
 use gen_constraint_input::NewGenConstraint;
 use input_data_setup_input::InputDataSetupUpdate;
-use job_status::JobStatus;
 use juniper::{
     graphql_object, Context, EmptySubscription, FieldResult, GraphQLInputObject, GraphQLObject,
     GraphQLUnion, Nullable, RootNode,
@@ -35,11 +36,10 @@ use process_input::NewProcess;
 use risk_input::NewRisk;
 use state_input::{StateInput, StateUpdate};
 use std::ops::DerefMut;
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 use time_line_input::TimeLineUpdate;
 use tokio::sync::mpsc;
-use tokio::sync::watch;
-use tokio::sync::Semaphore;
+use tokio::sync::Mutex;
 use topology_input::NewTopology;
 
 #[derive(Debug, GraphQLObject)]
@@ -97,19 +97,8 @@ enum SettingsResult {
 }
 
 #[derive(GraphQLObject)]
-struct StartOptimizationOutput {
-    job_id: i32,
-}
-
-#[derive(GraphQLObject)]
 struct StartOptimizationError {
     message: String,
-}
-
-#[derive(GraphQLUnion)]
-enum StartOptimizationResult {
-    Ok(StartOptimizationOutput),
-    Err(StartOptimizationError),
 }
 
 #[derive(GraphQLObject)]
@@ -140,31 +129,25 @@ impl MaybeError {
 
 pub struct HerttaContext {
     settings: Arc<Mutex<Settings>>,
+    job_store: JobStore,
     model: Arc<Mutex<Model>>,
-    job_id: Arc<Mutex<i32>>,
-    optimize_permit: Arc<Semaphore>,
-    tx_optimize: mpsc::Sender<(i32, OptimizationTask)>,
-    rx_state: watch::Receiver<OptimizationState>,
+    job_sender: mpsc::Sender<NewJob>,
 }
 
 impl Context for HerttaContext {}
 
 impl HerttaContext {
     pub fn new(
-        settings: &Arc<Mutex<Settings>>,
-        model: &Arc<Mutex<Model>>,
-        job_id: &Arc<Mutex<i32>>,
-        optimize_permit: &Arc<Semaphore>,
-        tx_optimize: mpsc::Sender<(i32, OptimizationTask)>,
-        rx_state: watch::Receiver<OptimizationState>,
+        settings: Arc<Mutex<Settings>>,
+        job_store: JobStore,
+        model: Arc<Mutex<Model>>,
+        job_sender: mpsc::Sender<NewJob>,
     ) -> Self {
         HerttaContext {
-            settings: Arc::clone(settings),
-            model: Arc::clone(model),
-            job_id: Arc::clone(job_id),
-            optimize_permit: Arc::clone(optimize_permit),
-            tx_optimize,
-            rx_state,
+            settings,
+            job_store,
+            model,
+            job_sender,
         }
     }
 
@@ -181,16 +164,19 @@ impl Query {
     fn api_version() -> &'static str {
         "0.10.0"
     }
-    fn settings(context: &HerttaContext) -> FieldResult<Settings> {
-        let settings = context.settings.lock().unwrap();
+    async fn settings(context: &HerttaContext) -> FieldResult<Settings> {
+        let settings = context.settings.lock().await;
         Ok(settings.clone())
     }
-    fn model(context: &HerttaContext) -> FieldResult<Model> {
-        let model = context.model.lock().unwrap();
+    async fn model(context: &HerttaContext) -> FieldResult<Model> {
+        let model = context.model.lock().await;
         Ok(model.clone())
     }
-    fn gen_constraint(name: String, context: &HerttaContext) -> FieldResult<BaseGenConstraint> {
-        let model = context.model.lock().unwrap();
+    async fn gen_constraint(
+        name: String,
+        context: &HerttaContext,
+    ) -> FieldResult<BaseGenConstraint> {
+        let model = context.model.lock().await;
         model
             .input_data
             .gen_constraints
@@ -199,8 +185,8 @@ impl Query {
             .map(|g| g.clone())
             .ok_or_else(|| "no such generic constraint".into())
     }
-    fn node_group(name: String, context: &HerttaContext) -> FieldResult<NodeGroup> {
-        let model = context.model.lock().unwrap();
+    async fn node_group(name: String, context: &HerttaContext) -> FieldResult<NodeGroup> {
+        let model = context.model.lock().await;
         model
             .input_data
             .node_groups
@@ -209,8 +195,8 @@ impl Query {
             .map(|g| g.clone())
             .ok_or_else(|| "no such group".into())
     }
-    fn nodes_in_group(name: String, context: &HerttaContext) -> FieldResult<Vec<BaseNode>> {
-        let model = context.model.lock().unwrap();
+    async fn nodes_in_group(name: String, context: &HerttaContext) -> FieldResult<Vec<BaseNode>> {
+        let model = context.model.lock().await;
         group_members(
             &model.input_data.node_groups,
             &name,
@@ -218,8 +204,8 @@ impl Query {
         )
         .map_err(|error| error.into())
     }
-    fn process_group(name: String, context: &HerttaContext) -> FieldResult<ProcessGroup> {
-        let model = context.model.lock().unwrap();
+    async fn process_group(name: String, context: &HerttaContext) -> FieldResult<ProcessGroup> {
+        let model = context.model.lock().await;
         model
             .input_data
             .process_groups
@@ -228,8 +214,11 @@ impl Query {
             .map(|g| g.clone())
             .ok_or_else(|| "no such group".into())
     }
-    fn processes_in_group(name: String, context: &HerttaContext) -> FieldResult<Vec<BaseProcess>> {
-        let model = context.model.lock().unwrap();
+    async fn processes_in_group(
+        name: String,
+        context: &HerttaContext,
+    ) -> FieldResult<Vec<BaseProcess>> {
+        let model = context.model.lock().await;
         group_members(
             &model.input_data.process_groups,
             &name,
@@ -237,8 +226,8 @@ impl Query {
         )
         .map_err(|error| error.into())
     }
-    fn market(name: String, context: &HerttaContext) -> FieldResult<BaseMarket> {
-        let model = context.model.lock().unwrap();
+    async fn market(name: String, context: &HerttaContext) -> FieldResult<BaseMarket> {
+        let model = context.model.lock().await;
         model
             .input_data
             .markets
@@ -247,8 +236,8 @@ impl Query {
             .map(|m| m.clone())
             .ok_or_else(|| "no such market".into())
     }
-    fn node(name: String, context: &HerttaContext) -> FieldResult<BaseNode> {
-        let model = context.model.lock().unwrap();
+    async fn node(name: String, context: &HerttaContext) -> FieldResult<BaseNode> {
+        let model = context.model.lock().await;
         model
             .input_data
             .nodes
@@ -258,8 +247,8 @@ impl Query {
             .ok_or_else(|| "no such node".into())
     }
     #[graphql(description = "Return all groups the given node is member of.")]
-    fn groups_for_node(name: String, context: &HerttaContext) -> FieldResult<Vec<NodeGroup>> {
-        let model = context.model.lock().unwrap();
+    async fn groups_for_node(name: String, context: &HerttaContext) -> FieldResult<Vec<NodeGroup>> {
+        let model = context.model.lock().await;
         member_groups(
             &model.input_data.nodes,
             &name,
@@ -267,12 +256,12 @@ impl Query {
         )
         .map_err(|error| error.into())
     }
-    fn node_diffusion(
+    async fn node_diffusion(
         from_node: String,
         to_node: String,
         context: &HerttaContext,
     ) -> FieldResult<BaseNodeDiffusion> {
-        let model = context.model.lock().unwrap();
+        let model = context.model.lock().await;
         model
             .input_data
             .node_diffusion
@@ -282,8 +271,11 @@ impl Query {
             .ok_or_else(|| "no such node diffusion".into())
     }
     #[graphql(description = "Return all groups the given process is member of.")]
-    fn groups_for_process(name: String, context: &HerttaContext) -> FieldResult<Vec<ProcessGroup>> {
-        let model = context.model.lock().unwrap();
+    async fn groups_for_process(
+        name: String,
+        context: &HerttaContext,
+    ) -> FieldResult<Vec<ProcessGroup>> {
+        let model = context.model.lock().await;
         member_groups(
             &model.input_data.processes,
             &name,
@@ -291,8 +283,8 @@ impl Query {
         )
         .map_err(|error| error.into())
     }
-    fn process(name: String, context: &HerttaContext) -> FieldResult<BaseProcess> {
-        let model = context.model.lock().unwrap();
+    async fn process(name: String, context: &HerttaContext) -> FieldResult<BaseProcess> {
+        let model = context.model.lock().await;
         model
             .input_data
             .processes
@@ -301,8 +293,8 @@ impl Query {
             .map(|p| p.clone())
             .ok_or_else(|| "no such process".into())
     }
-    fn scenario(name: String, context: &HerttaContext) -> FieldResult<Scenario> {
-        let model = context.model.lock().unwrap();
+    async fn scenario(name: String, context: &HerttaContext) -> FieldResult<Scenario> {
+        let model = context.model.lock().await;
         model
             .input_data
             .scenarios
@@ -311,30 +303,33 @@ impl Query {
             .map(|s| s.clone())
             .ok_or_else(|| "no such scenario".into())
     }
-    fn job_status(job_id: i32, context: &HerttaContext) -> FieldResult<JobStatus> {
-        match *context.rx_state.borrow() {
-            OptimizationState::Idle => Err("no such job".into()),
-            OptimizationState::InProgress(current_job_id) => {
-                if job_id == current_job_id {
-                    Ok(JobStatus::new_in_progress())
-                } else {
-                    Err("no such job".into())
-                }
+    async fn job_status(
+        job_id: i32,
+        context: &HerttaContext,
+    ) -> FieldResult<job_status::JobStatus> {
+        let status = context
+            .job_store
+            .job_status(job_id)
+            .await
+            .ok_or("no such job")?;
+        match *status {
+            jobs::JobStatus::Queued => Ok(job_status::JobStatus::new_queued()),
+            jobs::JobStatus::InProgress => Ok(job_status::JobStatus::new_in_progress()),
+            jobs::JobStatus::Failed(ref failure) => {
+                Ok(job_status::JobStatus::new_failed(failure.message().clone()))
             }
-            OptimizationState::Finished(current_job_id) => {
-                if job_id == current_job_id {
-                    Ok(JobStatus::new_finished())
-                } else {
-                    Err("no such job".into())
-                }
-            }
-            OptimizationState::Error(current_job_id, ref error) => {
-                if job_id == current_job_id {
-                    Ok(JobStatus::new_failed(error))
-                } else {
-                    Err("no such job".into())
-                }
-            }
+            jobs::JobStatus::Finished(..) => Ok(job_status::JobStatus::new_finished()),
+        }
+    }
+    async fn job_outcome(job_id: i32, context: &HerttaContext) -> FieldResult<JobOutcome> {
+        let job_status = context
+            .job_store
+            .job_status(job_id)
+            .await
+            .ok_or("no such job")?;
+        match *job_status {
+            JobStatus::Finished(ref outcome) => Ok(outcome.clone()),
+            _ => Err("job not finished".into()),
         }
     }
 }
@@ -393,93 +388,81 @@ pub struct Mutation;
 #[graphql_object]
 #[graphql(context = HerttaContext)]
 impl Mutation {
-    async fn start_optimization(context: &HerttaContext) -> StartOptimizationResult {
-        let current_job_id: i32;
-        {
-            let _permit = context.optimize_permit.acquire().await.unwrap();
-            if let OptimizationState::InProgress(..) = *context.rx_state.borrow() {
-                return StartOptimizationResult::Err(StartOptimizationError {
-                    message: "optimization already underway".to_string(),
-                });
-            }
-            {
-                let mut mutable_id = context.job_id.lock().unwrap();
-                current_job_id = *mutable_id;
-                *mutable_id += 1;
-            }
-            if context
-                .tx_optimize
-                .send((current_job_id, OptimizationTask::Start))
-                .await
-                .is_err()
-            {
-                return StartOptimizationResult::Err(StartOptimizationError {
-                    message: "failed to send task to event loop".to_string(),
-                });
-            }
-        }
-        StartOptimizationResult::Ok(StartOptimizationOutput {
-            job_id: current_job_id,
-        })
+    #[graphql(description = "Start optimization job. Return job ID.")]
+    async fn start_optimization(context: &HerttaContext) -> FieldResult<i32> {
+        let job_id =
+            event_loop::start_job(Job::Optimization, &context.job_store, &context.job_sender)
+                .await?;
+        Ok(job_id)
+    }
+    #[graphql(description = "Start weather forecast job. Return job ID.")]
+    async fn start_weather_forecast_fetch(context: &HerttaContext) -> FieldResult<i32> {
+        let job_id = event_loop::start_job(
+            Job::WeatherForecast,
+            &context.job_store,
+            &context.job_sender,
+        )
+        .await?;
+        Ok(job_id)
     }
 
     #[graphql(description = "Update model's time line.")]
-    fn update_time_line(
+    async fn update_time_line(
         time_line_input: TimeLineUpdate,
         context: &HerttaContext,
     ) -> ValidationErrors {
-        let mut model = context.model.lock().unwrap();
+        let mut model = context.model.lock().await;
         time_line_input::update_time_line(time_line_input, &mut model.time_line)
     }
 
     #[graphql(description = "Create new scenario.")]
-    fn create_scenario(name: String, weight: f64, context: &HerttaContext) -> MaybeError {
-        let mut model = context.model.lock().unwrap();
+    async fn create_scenario(name: String, weight: f64, context: &HerttaContext) -> MaybeError {
+        let mut model = context.model.lock().await;
         scenario_input::create_scenario(name, weight, &mut model.input_data.scenarios)
     }
 
     #[graphql(description = "Save the model on disk.")]
-    fn save_model(context: &HerttaContext) -> MaybeError {
+    async fn save_model(context: &HerttaContext) -> MaybeError {
         let file_path = model::make_model_file_path();
-        let model = context.model.lock().unwrap();
+        let model = context.model.lock().await;
         let result = model::write_model_to_file(&model, &file_path).err();
         MaybeError { error: result }
     }
 
     #[graphql(description = "Clear input data from model.")]
-    fn clear_input_data(context: &HerttaContext) -> MaybeError {
-        let mut lock_guard = context.model.lock();
-        let model = lock_guard.as_mut().unwrap();
+    async fn clear_input_data(context: &HerttaContext) -> MaybeError {
+        let mut lock_guard = context.model.lock().await;
+        let model = lock_guard.deref_mut();
         model.input_data = BaseInputData::default();
         MaybeError::new_ok()
     }
 
     #[graphql(description = "Update input data setup.")]
-    fn update_input_data_setup(
+    async fn update_input_data_setup(
         setup_update: InputDataSetupUpdate,
         context: &HerttaContext,
     ) -> ValidationErrors {
-        let mut model = context.model.lock().unwrap();
+        let mut model = context.model.lock().await;
         input_data_setup_input::update_input_data_setup(setup_update, &mut model.input_data.setup)
     }
 
     #[graphql(description = "Create new node group")]
-    fn create_node_group(name: String, context: &HerttaContext) -> MaybeError {
-        let mut model_ref = context.model.lock().unwrap();
+    async fn create_node_group(name: String, context: &HerttaContext) -> MaybeError {
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         group_input::create_node_group(name, &mut model.input_data.node_groups)
     }
 
     #[graphql(description = "Create new process group.")]
-    fn create_process_group(name: String, context: &HerttaContext) -> MaybeError {
-        let mut model_ref = context.model.lock().unwrap();
+    async fn create_process_group(name: String, context: &HerttaContext) -> MaybeError {
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         group_input::create_process_group(name, &mut model.input_data.process_groups)
     }
 
     #[graphql(description = "Create new process.")]
-    fn create_process(process: NewProcess, context: &HerttaContext) -> ValidationErrors {
-        let mut model_ref = context.model.lock().unwrap();
+    async fn create_process(process: NewProcess, context: &HerttaContext) -> ValidationErrors {
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         process_input::create_process(
             process,
@@ -489,12 +472,12 @@ impl Mutation {
     }
 
     #[graphql(description = "Add process to process group.")]
-    fn add_process_to_group(
+    async fn add_process_to_group(
         process_name: String,
         group_name: String,
         context: &HerttaContext,
     ) -> MaybeError {
-        let mut model_ref = context.model.lock().unwrap();
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         group_input::add_to_group(
             &process_name,
@@ -505,14 +488,14 @@ impl Mutation {
     }
 
     #[graphql(description = "Create new topology and add it to process.")]
-    fn create_topology(
+    async fn create_topology(
         topology: NewTopology,
         source_node_name: Option<String>,
         process_name: String,
         sink_node_name: Option<String>,
         context: &HerttaContext,
     ) -> ValidationErrors {
-        let mut model_ref = context.model.lock().unwrap();
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         topology_input::create_topology(
             process_name,
@@ -525,8 +508,8 @@ impl Mutation {
     }
 
     #[graphql(description = "Create new node.")]
-    fn create_node(node: NewNode, context: &HerttaContext) -> ValidationErrors {
-        let mut model_ref = context.model.lock().unwrap();
+    async fn create_node(node: NewNode, context: &HerttaContext) -> ValidationErrors {
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         node_input::create_node(
             node,
@@ -536,12 +519,12 @@ impl Mutation {
     }
 
     #[graphql(description = "Add node to node group.")]
-    fn add_node_to_group(
+    async fn add_node_to_group(
         node_name: String,
         group_name: String,
         context: &HerttaContext,
     ) -> MaybeError {
-        let mut model_ref = context.model.lock().unwrap();
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         group_input::add_to_group(
             &node_name,
@@ -552,35 +535,35 @@ impl Mutation {
     }
 
     #[graphql(description = "Set state for node. Null clears the state.")]
-    fn set_node_state(
+    async fn set_node_state(
         state: Option<StateInput>,
         node_name: String,
         context: &HerttaContext,
     ) -> ValidationErrors {
-        let mut model_ref = context.model.lock().unwrap();
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         state_input::set_state_for_node(&node_name, state, &mut model.input_data.nodes)
     }
 
     #[graphql(description = "Update state of a node. The state has to be set.")]
-    fn update_node_state(
+    async fn update_node_state(
         state: StateUpdate,
         node_name: String,
         context: &HerttaContext,
     ) -> ValidationErrors {
-        let mut model_ref = context.model.lock().unwrap();
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         state_input::update_state_in_node(state, node_name, &mut model.input_data.nodes)
     }
 
     #[graphql(description = "Create new diffusion between nodes.")]
-    fn create_node_diffusion(
+    async fn create_node_diffusion(
         from_node: String,
         to_node: String,
         coefficient: f64,
         context: &HerttaContext,
     ) -> ValidationErrors {
-        let mut model_ref = context.model.lock().unwrap();
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         node_diffusion_input::create_node_diffusion(
             from_node,
@@ -592,8 +575,8 @@ impl Mutation {
     }
 
     #[graphql(description = "Create new market.")]
-    fn create_market(market: NewMarket, context: &HerttaContext) -> ValidationErrors {
-        let mut model_ref = context.model.lock().unwrap();
+    async fn create_market(market: NewMarket, context: &HerttaContext) -> ValidationErrors {
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         market_input::create_market(
             market,
@@ -604,17 +587,17 @@ impl Mutation {
     }
 
     #[graphql(description = "Create new risk.")]
-    fn create_risk(risk: NewRisk, context: &HerttaContext) -> ValidationErrors {
-        let mut model = context.model.lock().unwrap();
+    async fn create_risk(risk: NewRisk, context: &HerttaContext) -> ValidationErrors {
+        let mut model = context.model.lock().await;
         risk_input::create_risk(risk, &mut model.input_data.risk)
     }
 
     #[graphql(description = "Create new generic constraint.")]
-    fn create_gen_constraint(
+    async fn create_gen_constraint(
         constraint: NewGenConstraint,
         context: &HerttaContext,
     ) -> ValidationErrors {
-        let mut model = context.model.lock().unwrap();
+        let mut model = context.model.lock().await;
         gen_constraint_input::create_gen_constraint(
             constraint,
             &mut model.input_data.gen_constraints,
@@ -622,14 +605,14 @@ impl Mutation {
     }
 
     #[graphql(description = "Create new flow constraint factor and add it to generic constraint.")]
-    fn create_flow_con_factor(
+    async fn create_flow_con_factor(
         factor: f64,
         constraint_name: String,
         process_name: String,
         source_or_sink_node_name: String,
         context: &HerttaContext,
     ) -> ValidationErrors {
-        let mut model_ref = context.model.lock().unwrap();
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         con_factor_input::create_flow_con_factor(
             factor,
@@ -642,13 +625,13 @@ impl Mutation {
     }
 
     #[graphql(description = "Create new state constraint factor and add it to generic constraint.")]
-    fn create_state_con_factor(
+    async fn create_state_con_factor(
         factor: f64,
         constraint_name: String,
         node_name: String,
         context: &HerttaContext,
     ) -> ValidationErrors {
-        let mut model_ref = context.model.lock().unwrap();
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         con_factor_input::create_state_con_factor(
             factor,
@@ -662,13 +645,13 @@ impl Mutation {
     #[graphql(
         description = "Create new online constraint factor and add it to generic constraint."
     )]
-    fn create_online_con_factor(
+    async fn create_online_con_factor(
         factor: f64,
         constraint_name: String,
         process_name: String,
         context: &HerttaContext,
     ) -> ValidationErrors {
-        let mut model_ref = context.model.lock().unwrap();
+        let mut model_ref = context.model.lock().await;
         let model = model_ref.deref_mut();
         con_factor_input::create_online_con_factor(
             factor,
@@ -679,9 +662,12 @@ impl Mutation {
         )
     }
 
-    fn update_settings(settings_input: SettingsInput, context: &HerttaContext) -> SettingsResult {
+    async fn update_settings(
+        settings_input: SettingsInput,
+        context: &HerttaContext,
+    ) -> SettingsResult {
         let errors = Vec::new();
-        let mut settings = context.settings.lock().unwrap();
+        let mut settings = context.settings.lock().await;
         match settings_input.location {
             Nullable::Some(ref location_input) => update_location(location_input, &mut settings),
             Nullable::ExplicitNull => settings.location = None,
@@ -714,22 +700,13 @@ mod tests {
 
     fn default_context() -> HerttaContext {
         let settings = Arc::new(Mutex::new(Settings::default()));
+        let job_store = JobStore::default();
         let model = Arc::new(Mutex::new(Model::default()));
-        let job_id = Arc::new(Mutex::new(0));
-        let optimize_permit = Arc::new(Semaphore::new(1));
-        let (tx_optimize, _rx_optimize) = mpsc::channel::<(i32, OptimizationTask)>(1);
-        let (_tx_state, rx_state) = watch::channel::<OptimizationState>(OptimizationState::Idle);
-        HerttaContext::new(
-            &settings,
-            &model,
-            &job_id,
-            &optimize_permit,
-            tx_optimize,
-            rx_state,
-        )
+        let (tx_optimize, _rx_optimize) = mpsc::channel::<NewJob>(1);
+        HerttaContext::new(settings, job_store, model, tx_optimize)
     }
-    #[test]
-    fn update_location_in_settings() {
+    #[tokio::test]
+    async fn update_location_in_settings() {
         let context = default_context();
         let mut input = SettingsInput::default();
         let location_input = LocationInput {
@@ -737,7 +714,7 @@ mod tests {
             place: Some("Akun puoti".to_string()),
         };
         input.location = Nullable::Some(location_input);
-        let output = Mutation::update_settings(input, &context);
+        let output = Mutation::update_settings(input, &context).await;
         let location_output = match output {
             SettingsResult::Ok(settings) => settings.location.expect("location should be there"),
             SettingsResult::Err(..) => panic!("setting location should not fail"),
@@ -745,7 +722,7 @@ mod tests {
         assert_eq!(location_output.country, "Puurtila".to_string());
         assert_eq!(location_output.place, "Akun puoti".to_string());
         {
-            let settings = context.settings.lock().unwrap();
+            let settings = context.settings.lock().await;
             assert_eq!(
                 settings
                     .location
